new-architecture/components/data-provider-api/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pytest==7.4.2
pytest-flask==1.3.0
requests==2.31.0
# OPAL External Data Sources dependencies
PyJWT==2.8.0
cryptography==41.0.7
# PostgreSQL dependencies
psycopg2-binary==2.9.7
